cast uint16 explicitly to int
shift[j] is of dtype uint16 and does have a value of 0 for certain
iterations. a subtraction of 1 can thus cause an overflow depending on
the numpy version.

in numpy < 2.0, shift[j] is automatically promoted to an int32 prior
to the subtraction and thus no overflow occurs.
(i.e. pyqtgraph was relying on this behavior)

in numpy >= 2.0, shift[j] remains as an uint16 and an overflow occurs.