fix GLViewWidget::renderToArray clobbered by QPainter
This bug is triggered when GLTextItem and/or GLGradientLegendItem are used.

QPainter rebinds the current framebuffer object. The unnecessary call to
glFramebufferTexture2D within the inner loop then wrongly acts upon this new
binding, which messes up the global state.

changed to use glReadPixels to be compatible with OpenGL ES